+++ src/app/mypage/account/_compornent/SplitArea/index.module.scss
@@ -1,11 +1,25 @@
+@use '@/app/_styles/mixin.module.scss' as *;
+
 .SplitArea {
   display: flex;
   justify-content: space-between;
+  flex-wrap: wrap;
   gap: 2rem;
+  @include mq(md) {
+    justify-content: center;
+    gap: 0;
+  }
   .InputArea {
     width: 60%;
+    @include mq(md) {
+      order: 2;
+      width: 100%;
+    }
     .Label {
       margin-bottom: 3rem;
+      @include mq(md) {
+        margin-bottom: 0;
+      }
       label {
         display: block;
         margin-bottom: 0.3em;
@@ -28,11 +42,23 @@
   .ImageArea {
     position: relative;
     width: 35%;
+    @include mq(md) {
+      order: 1;
+      margin-bottom: 2rem;
+      width: 30%;
+    }
+    @include mq(sm) {
+      width: 50%;
+    }
     .imageBox {
       position: absolute;
       top: 5%;
       right: 5%;
       z-index: 1;
+      @include mq(md) {
+        top: 0;
+        right: 0;
+      }
       button {
         background-color: #46A997;
         color: #fff;
